+++ docs/hazmat/primitives/padding.rst
@@ -44,17 +44,15 @@ multiple of the block size.
     .. method:: padder()
 
         :returns: A padding
-            :class:`~cryptography.hazmat.primitives.interfaces.PaddingContext`
+            :class:`~cryptography.hazmat.primitives.padding.PaddingContext`
             provider.
 
     .. method:: unpadder()
 
         :returns: An unpadding
-            :class:`~cryptography.hazmat.primitives.interfaces.PaddingContext`
+            :class:`~cryptography.hazmat.primitives.padding.PaddingContext`
             provider.
 
-
-.. currentmodule:: cryptography.hazmat.primitives.interfaces
 
 .. class:: PaddingContext
